+++ src/util/syserror.cpp
@@ -12,6 +12,12 @@
 #include <cstring>
 #include <string>
 
+#if defined(WIN32)
+#include <windows.h>
+#include <locale>
+#include <codecvt>
+#endif
+
 std::string SysErrorString(int err)
 {
     char buf[1024];
@@ -33,3 +39,21 @@ std::string SysErrorString(int err)
         return strprintf("Unknown error (%d)", err);
     }
 }
+
+#if defined(WIN32)
+std::string Win32ErrorString(int err)
+{
+    wchar_t buf[256];
+    buf[0] = 0;
+    if(FormatMessageW(FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS | FORMAT_MESSAGE_MAX_WIDTH_MASK,
+                       nullptr, err, MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT),
+                       buf, ARRAYSIZE(buf), nullptr))
+    {
+        return strprintf("%s (%d)", std::wstring_convert<std::codecvt_utf8_utf16<wchar_t>,wchar_t>().to_bytes(buf), err);
+    }
+    else
+    {
+        return strprintf("Unknown error (%d)", err);
+    }
+}
+#endif
